Updated mah schtuff

+++ src/twophasecommit/TransactionManager.java
@@ -6,7 +6,7 @@ import java.net.Socket;
  */
 public class TransactionManager {
 
-    public void transferMoney(int amount, Socket fromServer, Socket toServer){
+    public void transferMoney(double amount, Socket fromServer, Socket toServer){
         CohortQueryThread transferMoneyToQuery = new CohortQueryThread(toServer, amount + "", 2);
         CohortQueryThread transferMoneyFromQuery = new CohortQueryThread(fromServer, "-" + amount, 2);
         transferMoneyFromQuery.start();
